+++ lib/generators/modulorails/docker/templates/config/initializers/0_redis.rb
@@ -1,0 +1,2 @@
+REDIS_URL = ENV.fetch('REDIS_URL', 'redis://redis:6379')
+REDIS_CLI = Redis.new(url: REDIS_URL)